requirements.txt
annotated-types==0.5.0
anyio==3.7.1
Brotli==1.1.0
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
fastapi==0.103.1
h11==0.14.0
idna==3.4
mutagen==1.47.0
pycryptodomex==3.18.0
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
websockets==11.0.3
yt-dlp==2023.7.6
